app/requirements.txt
amqp==5.1.1
asgiref==3.5.2
async-timeout==4.0.2
billiard==3.6.4.0
celery==5.2.7
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
Django==4.1.4
django-filter==22.1
djangorestframework==3.14.0
flower==1.2.0
gunicorn==20.1.0
humanize==4.4.0
icmplib==3.0.3
install==1.3.5
kombu==5.2.4
Markdown==3.4.1
prometheus-client==0.15.0
prompt-toolkit==3.0.36
psycopg2-binary==2.9.1
python-dotenv==0.21.0
pytz==2022.6
redis==4.4.0
six==1.16.0
sqlparse==0.4.3
tornado==6.2
vine==5.0.0
wcwidth==0.2.5
